server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.0
diffusers==0.24.0
accelerate==0.24.1
pillow==10.1.0